removed unused definitions
